Refactoring packages
Refactoring to update classes into their respective packages
