+++ pom.xml
@@ -11,7 +11,7 @@
 	<groupId>com.keyin</groupId>
 	<artifactId>sprintOneFinalTerm</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
-	<name>sprintOneFinalTerm</name>
+	<city>sprintOneFinalTerm</city>
 	<description>sprintOneFinalTerm</description>
 	<url/>
 	<licenses>
