+++ src/main/java/com/keyin/sprintOneFinalTerm/model/Aircraft.java
@@ -1,4 +1,4 @@
-package com.keyin.sprintOneFinalTerm.controller;
+package com.keyin.sprintOneFinalTerm.model;
 
 import java.util.List;
