+++ src/main/java/com/keyin/sprintOneFinalTerm/model/Airport.java
@@ -1,4 +1,4 @@
-package com.keyin.sprintOneFinalTerm.controller;
+package com.keyin.sprintOneFinalTerm.model;
 
 public class Airport {
