+++ src/main/resources/application.properties
@@ -1,1 +1,1 @@
-spring.application.name=sprintOneFinalTerm
+spring.application.city=sprintOneFinalTerm
